+++ app/controllers/users_controller.rb
@@ -1,6 +1,11 @@
 class UsersController < ApplicationController
 
+  assign_records_for User
+  before_filter :assign_records
+
+  before_filter :ensure_user!, :except => [:index, :new, :create]
   before_filter :authenticate_user!, :except => [:index, :new, :create, :show]
+  before_filter :authorize_change!,  :except => [:index, :new, :create, :show]
 
   def new
     @user = User.new
@@ -18,7 +23,7 @@ class UsersController < ApplicationController
       end
 
       # Redirect to stored location for user in session or to the home page.
-      redirect_to stored_location_for(:user) || root_url
+      redirect_to stored_location_for(:user) || @user
     else
       render :new
     end
@@ -37,7 +42,7 @@ class UsersController < ApplicationController
     @user = current_user
     if @user.update_attributes(params[:user])
       flash[:success] = 'Updated successfully'
-      redirect_to root_path
+      redirect_to @user
     else
       render :edit
     end
@@ -49,5 +54,31 @@ class UsersController < ApplicationController
     flash[:success] = t('flash.users.destroy.notice', :default => 'User was removed')
     redirect_to root_path
   end
-  
+
+  protected
+
+  #===[ Helpers ]=========================================================
+
+  # Is the current user allowed to change this record?
+  def can_change?
+    if @user_found == true
+      return(@user && current_user && @user == current_user)
+    else
+      # NOTE: This should never happen because routing should prevent access to
+      # this resource without a username, and #assign_records will prevent
+      # loading of user records that don't exist.
+      return false
+    end
+  end
+  helper_method :can_change?
+
+  #===[ Filters ]=========================================================
+
+  # Only allow owner to change this record, else redirect with an error.
+  def authorize_change!
+    unless can_change?
+      respond_with_forbidden("You must be be this user to change the profile.")
+    end
+  end
+
 end